refactor: Update loader to be under src and fix signal deferred mode (Simply had to yield for ancestry added events to fire)
