+++ src/loader/default.project.json
@@ -1,6 +1,6 @@
 {
   "name": "loader",
   "tree": {
-    "$path": "src2"
+    "$path": "src"
   }
 }
