+++ src/loader/src/Utils.lua
@@ -6,7 +6,7 @@
 local Utils = {}
 
 local function errorOnIndex(_, index)
-	error(("Bad index %q"):format(tostring(index)), 2)
+	error(string.format("Bad index %q", tostring(index)), 2)
 end
 
 local READ_ONLY_METATABLE = {
