--- src/loader/src2/LoaderUtils.lua
@@ -1,3 +1,0 @@
--- Literally here just so old loading code works
-
-error("Not implemented")
